DE-3297: 1. removed convert_to_ts_is8601() 2. merged test cases with test_datetime_parser.py
