+++ task_script_utils/datetime_parser/__init__.py
@@ -1,3 +1,2 @@
-from .convert_to_iso_format import convert_to_ts_iso8601
 from .parser import parse
 from .datetime_config import DatetimeConfig
